+++ R/plotFactor.r
@@ -41,7 +41,7 @@
 #' 
 #' res <- ANOVA_DDCt(x = data,
 #'   numOfFactors = 2,
-#'   numberOfrefGenes = 2,
+#'   numberOfrefGenes = 3,
 #'   block = "block",
 #'   mainFactor.column = 2,
 #'   p.adj = "none")